Code/ai-recruitment-system/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0